pyproject.toml
[project]
name = "detectiveboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "alembic>=1.13",
]
